+++ src/main/java/com/api/product/presenter/response/ProductPresenterResponse.java
@@ -2,5 +2,7 @@ package com.api.product.presenter.response;
 
 import lombok.Builder;
 
+import java.math.BigDecimal;
+
 @Builder
-public record ProductPresenterResponse(int id, String name, String sku, Double price) {}
+public record ProductPresenterResponse(int id, String name, String sku, BigDecimal price) {}
