+++ src/test/java/com/api/product/entrypoint/fixture/ProductControllerTestFixture.java
@@ -4,14 +4,16 @@ import com.api.product.core.domain.Product;
 import com.api.product.core.dto.ProductDto;
 import com.api.product.core.dto.UpdateProductDto;
 
+import java.math.BigDecimal;
+
 public class ProductControllerTestFixture {
 
   private static final String ID = "1";
   private static final String NAME = "Bola de Futebol";
   private static final String SKU = "BOLA-123-ABC";
-  private static final Double PRICE = 10.0;
+  private static final BigDecimal PRICE = BigDecimal.valueOf(10.0);
   private static final String NEW_NAME = "Bola de Futebol 2";
-  private static final Double NEW_PRICE = 20.0;
+  private static final BigDecimal NEW_PRICE = BigDecimal.valueOf(20.0);
 
   public static ProductDto validRequest() {
     return new ProductDto(NAME, SKU, PRICE);
